moved user interaction commands to /bin folder
