+++ bin/game_cli.rb
